refactor: drop hydra-js re-exports, dedupe getImageUrl, stable listing test
Three changes that go together:

1. **Re-exports removed from `@volto-hydra/hydra-js`.** The previous
   extraction commit (e885f89) kept a 32-line re-export block in
   hydra.src.js for back-compat, which defeated the bundle shrink:
   esbuild followed the re-export and inlined the helper code right
   back into the bundle. This commit drops the re-exports completely
   and updates every caller in the repo to import the moved helpers
   from `@volto-hydra/helpers` directly.

   **hydra.js bundle: 456.6kb → 439.0kb (-17.6kb)** — first actual
   reduction since the extraction started.

2. **getImageUrl moved to `@volto-hydra/helpers`.** It was duplicated
   in `docs/examples/examples/{react,svelte,vue,astro}/utils.js`.
   Now the canonical implementation lives in helpers as a pure
   `getImageUrl(value, apiUrl)`. Each example's utils.js shrinks to
   a 3-line wrapper that curries apiUrl from whatever that frontend
   keeps it in (window._API_URL for browser, '' for SSR).

3. **Listing test races against async item resolution.** The previous
   assertion read `items.count()` once after `items.first()` became
   visible — but more items were still rendering. Switched to a
   stable-count poll (`expect.poll` with two-consecutive-same-value
   stability check). Caught a flake on react.

# Files

Workspace doc-example apps (4):
  - vite.config.js / astro.config.mjs — added `$helpers` alias
  - main.{js,jsx} — split imports between `$hydra` and `$helpers`
  - examples/{svelte,react,vue,astro}/utils.js — wrapper around
    helpers.getImageUrl
  - examples/astro/ListingBlock.astro — `$hydra` → `$helpers`

Workspace Volto admin:
  - package.json — added `@volto-hydra/helpers` as workspace dep
  - 4 source files split imports between hydra-js and helpers

Standalone examples (NOT in workspace, alias-only):
  - nuxt-blog-starter/nuxt.config.ts — added `@hydra-js/helpers` alias
  - hydra-vue-f7/vite.config.js + vite.config.test.js — same
  - 5 component files split imports (block.vue, ListingBlock.vue,
    ContextNavigationBlock.vue across nuxt + f7)

Bridge unit tests:
  - ssrSafe.test.js, ploneFetchItems.test.js — switched import path
    from `./hydra.src.js` to `@volto-hydra/helpers` (the functions
    they test live there now)

Helpers package:
  - index.js — appended `getImageUrl(value, apiUrl)`

Test infrastructure:
  - BlockVerificationHelper.ts — listing check uses `expect.poll`
    with two-consecutive-same-count stability gate

# Verification

  - hydra-js jest:        54/54 pass
  - hydra.js bundle:      439.0kb (was 456.6kb)
  - doc-examples astro:   20/20 pass
  - doc-examples svelte:  20/20 pass
  - doc-examples react:   20/20 pass (listing test now stable)
  - doc-examples vue:     20/20 pass

Standalone examples (nuxt, vue-f7) verified at import-resolution
level — alias configs wired, no stale imports of moved names.
