+++ docs/examples/examples/astro/ListingBlock.astro
@@ -22,7 +22,7 @@
  * That matches the placeholder behaviour for tests that only assert
  * the outer wrapper exists.
  */
-import { expandListingBlocks, ploneFetchItems, contentPath } from '$hydra';
+import { expandListingBlocks, ploneFetchItems, contentPath } from '$helpers';
 import SummaryItemBlock from './SummaryItemBlock.astro';
 
 const { block } = Astro.props;
