+++ docs/examples/test-astro/astro.config.mjs
@@ -26,6 +26,7 @@ export default defineConfig({
     resolve: {
       alias: {
         '$hydra': path.resolve(__dirname, '../../../packages/hydra-js/hydra.src.js'),
+        '$helpers': path.resolve(__dirname, '../../../packages/helpers/index.js'),
         '$examples': path.resolve(__dirname, '../examples/astro'),
         '$schemas': path.resolve(__dirname, '../block-definitions.json'),
       },
