+++ docs/examples/test-astro/src/main.js
@@ -8,7 +8,8 @@
  * frontend (Astro, PHP, Django, Rails) — only the render endpoint
  * implementation differs per language.
  */
-import { initBridge, expandListingBlocks, ploneFetchItems, contentPath } from '$hydra';
+import { initBridge } from '$hydra';
+import { expandListingBlocks, ploneFetchItems, contentPath } from '$helpers';
 import docPageDefinitions from '$schemas';
 
 const docBlocksConfig = Object.fromEntries(
