+++ docs/examples/test-svelte/src/main.js
@@ -1,4 +1,5 @@
-import { initBridge, expandListingBlocks, ploneFetchItems, contentPath } from '$hydra';
+import { initBridge } from '$hydra';
+import { expandListingBlocks, ploneFetchItems, contentPath } from '$helpers';
 import docPageDefinitions from '$schemas';
 const docBlocksConfig = Object.fromEntries(
   Object.values(docPageDefinitions).flatMap(page => Object.entries(page.blocks))
