+++ docs/examples/test-svelte/vite.config.js
@@ -66,6 +66,7 @@ export default defineConfig({
     alias: {
       '$examples': path.resolve(__dirname, '../examples/svelte'),
       '$hydra': path.resolve(__dirname, '../../../packages/hydra-js/hydra.src.js'),
+      '$helpers': path.resolve(__dirname, '../../../packages/helpers/index.js'),
       '$schemas': path.resolve(__dirname, '../block-definitions.json'),
     },
   },
