+++ docs/examples/test-vue/vite.config.js
@@ -79,6 +79,7 @@ export default defineConfig({
     alias: {
       '$examples': path.resolve(__dirname, '../examples/vue'),
       '$hydra': path.resolve(__dirname, '../../../packages/hydra-js/hydra.src.js'),
+      '$helpers': path.resolve(__dirname, '../../../packages/helpers/index.js'),
       '$schemas': path.resolve(__dirname, '../block-definitions.json'),
     },
   },
