+++ examples/hydra-vue-f7/vite.config.js
@@ -5,6 +5,7 @@ const SRC_DIR = path.resolve(__dirname, './src');
 const PUBLIC_DIR = path.resolve(__dirname, './public');
 const BUILD_DIR = path.resolve(__dirname, './dist');
 const HYDRA_JS_DIR = path.resolve(__dirname, '../../packages/hydra-js');
+const HELPERS_DIR = path.resolve(__dirname, '../../packages/helpers');
 
 // For local HTTPS dev, use: pnpm exec vite --config vite.config.https.js
 export default {
@@ -32,6 +33,7 @@ export default {
     alias: {
       '@': SRC_DIR,
       '@hydra-js/hydra.js': path.resolve(HYDRA_JS_DIR, 'hydra.src.js'),
+      '@hydra-js/helpers': path.resolve(HELPERS_DIR, 'index.js'),
       '@hydra-js': HYDRA_JS_DIR,
     },
   },
