+++ packages/volto-hydra/package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@plone/components": "workspace:*",
+    "@volto-hydra/helpers": "workspace:*",
     "@volto-hydra/hydra-js": "workspace:*",
     "immer": "^10.0.0",
     "js-cookie": "^3.0.5",
